requirements.txt
Django==2.2.7
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
# mysqlclient==1.4.6
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
